moved flatten func to new helperFuunctions file
