+++ test/routes/mino/Mino.ts
@@ -8,12 +8,18 @@ export default class MinoRoute extends Route {
       method: 'get',
       path: 'mino',
       collections: ['mino', 'non_existant'],
+      middleware: [
+        (req, res, next) => {
+          console.log('mino middleware', req.headers);
+          next();
+        },
+      ],
     });
   }
 
   async run(args: IRouteArgs) {
     const { request, response } = args;
-    console.log(request.headers);
+    console.log('mino run', request.headers);
     return response.send('Mino!');
   }
 }
